+++ include/ccf/historical_queries_utils.h
@@ -41,8 +41,9 @@ namespace ccf::historical
     ccf::historical::StatePtr& state,
     AbstractStateCache& state_cache);
 
-  // Verifies CCF COSE receipt using the current network identity's certificate.
-  void verify_cose_receipt(
+  // Verifies CCF COSE receipt using the *current network* identity's
+  // certificate.
+  void verify_self_issued_receipt(
     const std::vector<uint8_t>& cose_receipt,
     std::shared_ptr<NetworkIdentitySubsystemInterface>
       network_identity_subsystem);